xml/parsing/manakai/pis-2.dat

#data
<!DOCTYPE a[
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x y?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
<?y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
]>
<?y?>
<a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <?y ?>
| <{} a>

#data
<!DOCTYPE a[
<?x ?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x ??y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ??y?>
| <{} a>

#data
<!DOCTYPE a[
<?x
#errors
1;1;no XML decl;;;s
2;4;no pic;;;m
2;4;unclosed internal subset;;;m
2;4;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<?x\u0020
#errors
1;1;no XML decl;;;s
2;5;no pic;;;m
2;5;unclosed internal subset;;;m
2;5;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?x y
#errors
1;1;no XML decl;;;s
2;6;no pic;;;m
2;6;unclosed internal subset;;;m
2;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?x ?
#errors
1;1;no XML decl;;;s
2;6;no pic;;;m
2;6;unclosed internal subset;;;m
2;6;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?
#errors
1;1;no XML decl;;;s
2;1;bare pio;;;m
2;3;unclosed internal subset;;;m
2;3;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?>
#errors
1;1;no XML decl;;;s
2;4;no pic;;;m
2;4;unclosed internal subset;;;m
2;4;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<?>
]><a/>
#errors
1;1;no XML decl;;;s
3;7;no pic;;;m
3;7;unclosed internal subset;;;m
3;7;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<??>
<?x?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;bare pio;;;m
3;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
| <{} a>

#data
<!DOCTYPE a[
<?x?>
%b;
<?y?>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:pi;;;w
3;1;pe not declared;;b;;m
3;1;stop processing;;;i
4;1;xml:dtd:pi;;;w
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>

#data
<?xml version="1.0" standalone="yes"?>
<!DOCTYPE a[
<?x?>
%b;
<?y?>
]><a/>
#errors
3;1;xml:dtd:pi;;;w
4;1;pe not declared;;b;;m
5;1;xml:dtd:pi;;;w
#xml-standalone
true
#document
| <!DOCTYPE a>
|   <?x ?>
|   <?y ?>
| <{} a>
